custom charts: Add core classes
A data model represents every plottable series coming from an
analysis. A series is described through a data descriptor which
contains an aspect and a source. The aspect tells information (e.g.
continuity) about the series. The source is the stream of data.

In order to make a chart from the plugin, an analysis must
extend the data model and create a descriptor for each plottable
series. Other patches will be coming as examples.

Change-Id: Ic93982ffd92f2eb582f9ad120ec9ba7fe87d2a7a
Signed-off-by: Gabriel-Andrew Pollo-Guilbert <gabrielpolloguilbert@gmail.com>