custom charts: Add core classes
A data provider provides all the informations needed to obtain
data in order to make a chart. It has a source that can be of any
object and a list of data descriptors that can map specific values
from a object.

In order to make a chart from the plugin, an analysis must
implement a data provider and create a data descriptor for each
data stream that can be used for plotting. Other patches will be
coming as examples.

Change-Id: Ic93982ffd92f2eb582f9ad120ec9ba7fe87d2a7a
Signed-off-by: Gabriel-Andrew Pollo-Guilbert <gabrielpolloguilbert@gmail.com>