custom charts: Add view for testing the plugin
This patch is not intended for merging. It  adds a new
view for testing the plugin. In order to make a chart,
a user must first open an analysis that support the
data model. All LAMI analyses should support it, as
well as the system call latency statistics analysis.

When an analysis is opened, its data model should be
visible in the chart maker of the view. Be warned that
the chart maker is outdated in this commit and has some
bugs. The "real" chart maker will be coming later.

Change-Id: Ia790b11083ef07b49953cf9f3578eeaea7785495
Signed-off-by: Gabriel-Andrew Pollo-Guilbert <gabrielpolloguilbert@gmail.com>